% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_editor_json.R
\name{parse_editor_json}
\alias{parse_editor_json}
\title{parses json produced by editorjs}
\usage{
parse_editor_json(out_json, ...)
}
\arguments{
\item{out_json}{object passed to jsonlite::fromJSON}

\item{...}{additional arguments passed to jsonlite::fromJSON}
}
\description{
parses json produced by editorjs
}
